+++ egg-info/history.rst
@@ -1,7 +1,55 @@
+zerobug: 2.0.0.1 (2022-09-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] travis_intsall_env: minor fixes
+
+
+
+z0lib: 2.0.0.1 (2022-09-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] set_pybin accept filename
+* [IMP] check_pythonpath removed
+
+
+travis_emulator: 2.0.0.1 (2022-09-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] travis: new improvements (-f -k switches)
+
+
+
+python_plus: 2.0.0.1 (2022-09-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] vem: new swith -d for Odoo dependencies path
+* [FIX] vem: create with best package list
+* [FIX] vem: install odoo/openerp
+
+
+
+clodoo: 2.0.0.1 (2022-09-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] list_requirements.py: new option -S for secure packages
+
+
+zerobug: 2.0.0 (2022-08-10)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [REF] Partial refactoring for shell scripts
+
+
 z0lib: 2.0.0 (2022-08-10)
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
 * [REF] Partial refactoring for shell scripts
+
+
+z0bug_odoo: 2.0.0 (2022-08-10)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [REF] Stable version
 
 
 wok_code: 2.0.0 (2022-08-10)
@@ -10,18 +58,35 @@ wok_code: 2.0.0 (2022-08-10)
 * [REF] Refactoring
 
 
-odoo_score: 1.0.8 (2022-07-30)
+travis_emulator: 2.0.0 (2022-08-10)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [REF] Partial refactoring for shell scripts
+
+
+python_plus: 2.0.0 (2022-08-10)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] Stable version
+
+
+odoo_score: 2.0.0 (2022-08-10)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-* [IMP] New command rename_odoo_module
+* [REF] Stable version
 
 
 
-clodoo: 1.0.7 (2022-07-30)
+lisa: 2.0.0 (2022-08-10)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [REF] Stable version
+
+
+clodoo: 2.0.0 (2022-08-10)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-* [FIX] clodoo: remote login
-* [IMP] Crypto.Cipher recognition
+* [REF] Stable version