+++ python_plus/egg-info/history.rst
@@ -3,6 +3,7 @@
 
 * [IMP] vem: new swith -d for Odoo dependencies path
 * [FIX] vem: create with best package list
+* [FIX] vem: install odoo/openerp
 
 
 2.0.0 (2022-08-10)